rename keytar Windows native module
Summary:
`process.platform` is used to require the native-module for keytar. On Windows this returns `win32` (No matter if 64bit or 32bit system). The native module was named `keytar-win.node`. This diff renames it to `keytar-win32.node`.

This is the code that requires the module from `user.js`:

```
keytar = global.electronRequire(
    path.join(
      isProduction() ? remote.app.getAppPath() : process.cwd(),
      'native-modules',
      `keytar-${process.platform}.node`,
    ),
  );
```

Reviewed By: priteshrnandgaonkar

Differential Revision: D14771646

fbshipit-source-id: cd48d25cb5921e07e472eb46636fa2599c8ff721