Write up FlipperClient API.
Summary:
Specifies the API for implementing your own Flipper Client.
I believe this provides everything you need to know to build a flipper-compliant client.
The only snag is that the current certificate exchange procedure would require a change to allow for new OS's, device types etc. It would be much better if we had a TLS-opt out so you don't need to go through all that rigmarole.

Reviewed By: passy

Differential Revision: D15150550

fbshipit-source-id: 2878dcc174807c3f86c0562660e8a3e43e5777b1