Redefine Js-Setup as JS Plugin Definition
Summary:
The purpose of this page wasn't that clear to me, though the content is pretty focused. I've changed it from JS setup to JS Plugin Definition, and reworded it a bit.
I think it makes sense if this specifies everything to do with plugins on the file system, and the next page talks about what you do in javascript.

Reviewed By: passy

Differential Revision: D15184829

fbshipit-source-id: 6335ebc8d4297ffcbf1ab9860b4f055fe5bbcf87